Refactor: Moved files in more specialized directories
