+++ src/client/components/PlaygroundRenderer.js
